手动上传简历的增删改查

+++ ruoyi-resume/src/main/java/com/ruoyi/resume/mapper/PerEducationMapper.java
@@ -49,7 +49,7 @@ public interface PerEducationMapper
      * @param id 简历教育经验ID
      * @return 结果
      */
-    public int deletePerEducationById(Long id);
+    public int deletePerEducationById(Integer id);
 
     /**
      * 批量删除简历教育经验
