+++ ruoyi-resume/src/main/java/com/ruoyi/resume/service/IPerWorkService.java
@@ -49,7 +49,7 @@ public interface IPerWorkService
      * @param ids 需要删除的简历录入工作经历ID
      * @return 结果
      */
-    public int deletePerWorkByIds(Long[] ids);
+    public int deletePerWorkByIds(Integer[] ids);
 
     /**
      * 删除简历录入工作经历信息
@@ -57,5 +57,5 @@ public interface IPerWorkService
      * @param id 简历录入工作经历ID
      * @return 结果
      */
-    public int deletePerWorkById(Long id);
+    public int deletePerWorkById(Integer id);
 }
